+++ contracts/debug/DebugStorageSlots.sol
@@ -136,7 +136,7 @@ contract DebugStorageSlots {
     uint32 minXP;
     uint32 diff;
     uint32 rate; // rate of output produced per hour (base 10) 1 decimal
-    uint16 xpPerHour;
+    uint24 xpPerHour;
     uint16 inputTokenId1;
     uint8 num1;
     uint16 inputTokenId2;
@@ -190,7 +190,7 @@ contract DebugStorageSlots {
     bool isAvailable;
     bool isDynamic;
     bool actionChoiceRequired; // If true, then the user must choose an action choice
-    uint16 xpPerHour;
+    uint24 xpPerHour;
     uint16 numSpawn; // Mostly for combat, capped respawn rate for xp/drops
     uint32 minXP;
     uint16 handItemTokenIdRangeMin; // Inclusive
